Split dialogs into several files
