+++ src/ui/components/dialogs/mod.rs
@@ -1,0 +1,5 @@
+pub mod label_dialogs;
+pub mod project_dialogs;
+pub mod scroll_behavior;
+pub mod system_dialogs;
+pub mod task_dialogs;
