+++ src/ui/components/mod.rs
@@ -4,6 +4,7 @@ pub mod badge;
 
 // Component architecture
 pub mod dialog_component;
+pub mod dialogs;
 pub mod sidebar_component;
 pub mod task_list_component;
